refactor: move `actionForBy` to helpers
